inserito tiling cambiata verifica convergenza - ora usa il numero di punti swappati ed una soglia basata sul numero totale di punti

+++ Point.h
@@ -11,13 +11,13 @@ public:
 
 	int ID;
 	int numAttributes;
-	double* attributes;
+	float* attributes;
 
 	Point(){
 
 	}
 
-	Point(int id, int numAttribs, double* attribs){
+	Point(int id, int numAttribs, float* attribs){
 		ID = id;
 		numAttributes = numAttribs;
 		attributes = attribs;
